[api/utils] structural changes
