+++ api/main.py
@@ -1,7 +1,7 @@
 from flask import Flask
 from flask_restful import Api
 
-from utils.api import add_app_resources_from
+from api.tools import add_app_resources_from
 
 
 # TODO middleware - login required for some apps/paths.
@@ -14,3 +14,5 @@ if __name__ == "__main__":
     add_app_resources_from(api, "api.app")
 
     app.run(debug=True)
+
+    Exception
